requirements.txt
appdirs==1.4.3
asgiref==3.2.7
Babel==2.8.0
certifi==2019.11.28
codicefiscale==0.9
distlib==0.3.0
dj-database-url==0.5.0
Django==3.0.4
django-crispy-forms==1.9.0
django-extensions==2.2.8
django-localflavor==3.0.1
django-phonenumber-field==4.0.0
django-resized==0.3.9
djangorestframework==3.11.0
filelock==3.0.12
gunicorn==20.0.4
importlib==1.0.4
importlib-metadata==1.6.0
importlib-resources==1.4.0
phonenumbers==8.12.1
Pillow==7.1.1
pipenv==2018.11.26
psycopg2-binary==2.8.4
python-stdnum==1.13
pytz==2019.3
six==1.14.0
sqlparse==0.3.1
Unidecode==1.1.1
utils==1.0.1
virtualenv==20.0.15
virtualenv-clone==0.5.4
Werkzeug==1.0.0
whitenoise==5.0.1
zipp==3.1.0
django-modeladmin-reorder
django-cleanup
